First steps towards Seat/Room/Time selection
* src/main/java/is/hi/travel_planer/control/PackageController.java:
(bookTour, bookPackage): Use the new TourTime class instead of a pair
of LocalDate LocalTime arguments.
* src/main/java/is/hi/travel_planer/model/TourTime.java (TourTime):
Create a new class to manage a specific time slot for a DayTour, since
DayTourTimes seems unsuited for this purpose.
* src/main/java/is/hi/travel_planer/view/DetailSelectionController.java
(DetailSelectionController): Create a new class to manage the
selection of Seats, Rooms and TourTime.
* src/main/java/is/hi/travel_planer/view/PackageSelectionController.java:
(handleNext): Use DetailSelectionController and DetailSelection.fxml
instead of the uncommitted DetailSelector and DetailSelector.fxml
* src/main/resources/fxml/DetailSelection.fxml: Create